Add yomu & spectral
