--- src/pt/sssscanlator/build.gradle
@@ -1,13 +1,0 @@
-ext {
-    extName = 'SSSScanlator'
-    extClass = '.SSSScanlator'
-    themePkg = 'mangathemesia'
-    baseUrl = 'https://ssstoons.com'
-    overrideVersionCode = 4
-}
-
-apply from: "$rootDir/common.gradle"
-
-dependencies {
-    implementation project(':lib:randomua')
-}
