--- src/pt/tsukimangas/build.gradle
@@ -1,8 +1,0 @@
-ext {
-    extName = 'Tsuki Mangás'
-    extClass = '.TsukiMangas'
-    extVersionCode = 7
-    isNsfw = true
-}
-
-apply from: "$rootDir/common.gradle"
